Split data module.
