+++ src/continuity/data/shape.py
@@ -1,3 +1,7 @@
+"""
+`continuity.data.shape`
+"""
+
 from dataclasses import dataclass
